+++ tests/test_customer_router.py
@@ -1,17 +1,3 @@
-import pytest
-from unittest.mock import MagicMock
-from main import app
-from fastapi.testclient import TestClient
-from routes.customer_router import get_customer_service
-
-
-@pytest.fixture
-def router():
-    mock_service = MagicMock()
-    app.dependency_overrides[get_customer_service] = lambda: mock_service
-    router = TestClient(app)
-    return mock_service, router
-
 
 def test_route_delete_customer(router):
     mock_service, client = router
